refactor(monitoring): split REST composition providers
